GSYE-375: Move MarketResultsAggregator to aggregate_results module
